DRY shared content-security methods into UploaderContentSecurity module + sync delta specs
