+++ app/helpers/camaleon_cms/uploader_helper.rb
@@ -5,6 +5,7 @@ require 'tempfile'
 
 module CamaleonCms
   module UploaderHelper
+    include UploaderContentSecurity
     include ActionView::Helpers::NumberHelper
     include CamaleonCms::CamaleonHelper
 
@@ -428,45 +429,6 @@ module CamaleonCms
       { file: tempfile, error: nil }
     rescue StandardError => e
       { error: "Unable to download remote file: #{ERB::Util.html_escape(e.message)}" }
-    end
-
-    def svg_upload?(uploaded_io)
-      file_path = if uploaded_io.is_a?(ActionDispatch::Http::UploadedFile)
-                    uploaded_io.original_filename
-                  else
-                    uploaded_io.path
-                  end
-      file_path&.end_with?('.svg')
-    end
-
-    def svg_content_unsafe?(uploaded_io)
-      file = uploaded_io.is_a?(ActionDispatch::Http::UploadedFile) ? uploaded_io.tempfile : uploaded_io
-      content = file.read
-      file.rewind if file.respond_to?(:rewind)
-      CamaleonCms::SvgContentChecker.unsafe?(content)
-    end
-
-    def file_content_unsafe?(uploaded_io)
-      file = uploaded_io.is_a?(ActionDispatch::Http::UploadedFile) ? uploaded_io.tempfile : uploaded_io
-      return nil if svg_upload?(uploaded_io)
-
-      file_content_unsafe = nil
-
-      file.set_encoding(Encoding::BINARY) if file.respond_to?(:binmode) && file.respond_to?(:set_encoding)
-
-      # Read the file for pattern scanning, then rewind so subsequent consumers
-      # (e.g. upload handlers) can read the full content. Failing to rewind
-      # resulted in 0-byte uploads when the file was a Tempfile (see report).
-      file_content = file.read
-      file.rewind if file.respond_to?(:rewind)
-      CamaleonCms::ContentSecurity::SUSPICIOUS_PATTERNS.each do |pattern|
-        if file_content&.match?(pattern)
-          Rails.logger.info { "Potentially malicious content found: #{pattern.inspect}" }
-          break file_content_unsafe = pattern.inspect
-        end
-      end
-
-      file_content_unsafe
     end
 
     # helper for resize and crop method
