+++ lib/camaleon_cms.rb
@@ -3,6 +3,7 @@ require 'camaleon_cms/version'
 require 'camaleon_cms/content_security'
 require 'camaleon_cms/svg_content_checker'
 require 'camaleon_cms/media_security_headers'
+require 'camaleon_cms/uploader_content_security'
 
 module CamaleonCms
 end
